requirements.txt
Flask
request
lxml
